QuestionモデルとAnswerモデルへのDRY原則適用
