+++ app/Answer.php
@@ -7,6 +7,8 @@ use Parsedown;
 
 class Answer extends Model
 {
+    use VotableTrait;
+
     protected $fillable = ['body', 'user_id'];
 
     /**
@@ -91,35 +93,5 @@ class Answer extends Model
     public function isBest()
     {
         return $this->id === $this->question->best_answer_id;
-    }
-
-    /**
-     * このvotableを付けた全userの取得
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphToMany
-     */
-    public function votes()
-    {
-        return $this->morphToMany(User::class, 'votable');
-    }
-
-    /**
-     * answerの投票に1加算
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphToMany
-     */
-    public function upVotes()
-    {
-        return $this->votes()->wherePivot('vote', 1);
-    }
-
-    /**
-     * answerの投票に1減算
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphToMany
-     */
-    public function downVotes()
-    {
-        return $this->votes()->wherePivot('vote', -1);
     }
 }
