+++ app/User.php
@@ -119,22 +119,11 @@ class User extends Authenticatable
     {
         $voteQuestions = $this->voteQuestions();
 
-        if ($voteQuestions->where('votable_id', $question->id)->exists()) {
-            $voteQuestions->updateExistingPivot($question, ['vote' => $vote]);
-        } else {
-            $voteQuestions->attach($question, ['vote' => $vote]);
-        }
-
-        $question->load('votes');
-        $downVotes = (int) $question->downVotes()->sum('vote');
-        $upVotes = (int) $question->upVotes()->sum('vote');
-
-        $question->votes_count = $upVotes + $downVotes;
-        $question->save();
+        $this->_vote($voteQuestions, $question, $vote);
     }
 
     /**
-     * answer 投票機能
+     * answerの投票機能
      *
      * @param Question $question
      * @param $vote
@@ -143,17 +132,29 @@ class User extends Authenticatable
     {
         $voteAnswers = $this->voteAnswers();
 
-        if ($voteAnswers->where('votable_id', $answer->id)->exists()) {
-            $voteAnswers->updateExistingPivot($answer, ['vote' => $vote]);
+        $this->_vote($voteAnswers, $answer, $vote);
+    }
+
+    /**
+     * 投票機能
+     *
+     * @param $relationship
+     * @param $model
+     * @param $vote
+     */
+    private function _vote($relationship, $model, $vote)
+    {
+        if ($relationship->where('votable_id', $model->id)->exists()) {
+            $relationship->updateExistingPivot($model, ['vote' => $vote]);
         } else {
-            $voteAnswers->attach($answer, ['vote' => $vote]);
+            $relationship->attach($model, ['vote' => $vote]);
         }
 
-        $answer->load('votes');
-        $downVotes = (int) $answer->downVotes()->sum('vote');
-        $upVotes = (int) $answer->upVotes()->sum('vote');
+        $model->load('votes');
+        $downVotes = (int) $model->downVotes()->sum('vote');
+        $upVotes = (int) $model->upVotes()->sum('vote');
 
-        $answer->votes_count = $upVotes + $downVotes;
-        $answer->save();
+        $model->votes_count = $upVotes + $downVotes;
+        $model->save();
     }
 }
